测试断言优化与底层实现细节修正
本次提交主要包括以下内容：

- 测试代码断言由 IsTrue/== 替换为 AreEqual/HasCount/IsGreaterThan，提升可读性和准确性。
- DateTime 相关扩展方法统一调用 DateTimeUtil 静态方法，增强一致性。
- TSIDUtil 生成逻辑修正，避免符号扩展带来的潜在问题。
- HttpClientBuilder、RedisCacheProvider、TempFileUtil 增加预留字段并屏蔽未使用警告，便于后续扩展。
- TlsUtil 不安全协议判断方式调整，兼容更多平台。
- MSTest 配置项完善，抑制部分测试警告。
- 其他小幅代码清理与死代码修正。

整体提升了代码的健壮性、可维护性和未来扩展能力。

+++ EasyTool.Core/IdentifierCategory/TSIDUtil.cs
@@ -61,8 +61,8 @@ namespace EasyTool.IdentifierCategory
                 _lastTimestamp = timestamp;
 
                 return ((timestamp & MaxTimestamp) << (NodeIdBits + SequenceBits))
-                       | ((long)_nodeId << SequenceBits)
-                       | _sequence;
+                       | ((long)(uint)_nodeId << SequenceBits)
+                       | (long)(uint)_sequence;
             }
         }
 
@@ -107,8 +107,8 @@ namespace EasyTool.IdentifierCategory
                 _lastTimestamp = timestamp;
 
                 return ((timestamp & MaxTimestamp) << (NodeIdBits + SequenceBits))
-                       | ((long)nodeId << SequenceBits)
-                       | _sequence;
+                       | ((long)(uint)nodeId << SequenceBits)
+                       | (long)(uint)_sequence;
             }
         }
